added reqs and updated demo
