+++ models/other_models/FoldingNet.py
@@ -1,6 +1,6 @@
 import torch
 import torch.nn as nn
-from .build import MODELS
+from ..build import MODELS
 from extensions.chamfer_dist import ChamferDistanceL2
 
 @MODELS.register_module()
