+++ requirements.txt
@@ -1,0 +1,20 @@
+easydict==1.12
+einops==0.7.0
+gradio==4.18.0
+h5py==3.10.0
+matplotlib==3.7.4
+numpy==1.24.4
+open3d==0.11.2
+Pillow==9.3.0
+plotly==5.18.0
+pytorch3d==0.6.2
+PyYAML==6.0.1
+setuptools==68.2.2
+structures==0.9.5
+tensorboardX==2.6.2.2
+termcolor==2.4.0
+timm==0.9.12
+torch==1.8.0
+torchvision==0.9.0
+transforms3d==0.4.1
+wandb==0.16.3
